+++ react/typings/global.d.ts
@@ -1,0 +1,9 @@
+interface Window extends Window {
+  _iub: Iubenda
+}
+
+interface Iubenda {
+  advertisingPreferencesButton: {
+    enableElement: (element: HTMLElement) => void
+  }
+}