+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Actively bind element.
+
 ## [0.0.2] - 2020-04-17
 
-## [0.0.1] - 2020-04-17
-
 ### Added
+
 - Intial release.